style: rename function names
